+++ Tests/Functional/EventListener/AutoRouteListenerTest.php
@@ -3,14 +3,13 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-
-namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Functional\Subscriber;
+namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Functional\EventListener;
 
 use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Functional\BaseTestCase;
 use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document\Blog;
@@ -28,14 +27,14 @@ class AutoRouteListenerTest extends BaseTestCase
 {
     protected function createBlog($withPosts = false)
     {
-        $blog = new Blog;
+        $blog = new Blog();
         $blog->path = '/test/test-blog';
         $blog->title = 'Unit testing blog';
 
         $this->getDm()->persist($blog);
 
         if ($withPosts) {
-            $post = new Post;
+            $post = new Post();
             $post->name = 'This is a post title';
             $post->title = 'This is a post title';
             $post->blog = $blog;
@@ -47,6 +46,10 @@ class AutoRouteListenerTest extends BaseTestCase
         $this->getDm()->clear();
     }
 
+    /**
+     * It should persist the blog document and create an auto route.
+     * It should set the defaults on the route.
+     */
     public function testPersistBlog()
     {
         $this->createBlog();
@@ -62,7 +65,12 @@ class AutoRouteListenerTest extends BaseTestCase
         $this->assertCount(1, $routes);
         $this->assertInstanceOf('Symfony\Cmf\Bundle\RoutingAutoBundle\Model\AutoRoute', $routes[0]);
         $this->assertEquals('unit-testing-blog', $routes[0]->getName());
-        $this->assertEquals(PhpcrOdmAdapter::TAG_NO_MULTILANG, $routes[0]->getAutoRouteTag());
+        $this->assertEquals(PhpcrOdmAdapter::TAG_NO_MULTILANG, $routes[0]->getLocale());
+        $this->assertEquals(array(
+            '_auto_route_tag' => 'no-multilang',
+            'type' => 'cmf_routing_auto.primary',
+            '_controller' => 'BlogController',
+        ), $routes[0]->getDefaults());
     }
 
     public function provideTestUpdateBlog()
@@ -207,6 +215,16 @@ class AutoRouteListenerTest extends BaseTestCase
                     'test/auto-route/articles/fr/bonjour-le-monde',
                     'test/auto-route/articles/de/gutentag',
                     'test/auto-route/articles/es/hola-todo-el-mundo',
+
+                    'test/auto-route/articles/en/hello-everybody-edit',
+                    'test/auto-route/articles/fr/bonjour-le-monde-edit',
+                    'test/auto-route/articles/de/gutentag-edit',
+                    'test/auto-route/articles/es/hola-todo-el-mundo-edit',
+
+                    'test/auto-route/articles/en/hello-everybody-review',
+                    'test/auto-route/articles/fr/bonjour-le-monde-review',
+                    'test/auto-route/articles/de/gutentag-review',
+                    'test/auto-route/articles/es/hola-todo-el-mundo-review',
                 ),
             ),
         );
@@ -217,7 +235,7 @@ class AutoRouteListenerTest extends BaseTestCase
      */
     public function testMultilangArticle($data, $expectedPaths)
     {
-        $article = new Article;
+        $article = new Article();
         $article->path = '/test/article-1';
         $this->getDm()->persist($article);
 
@@ -229,17 +247,17 @@ class AutoRouteListenerTest extends BaseTestCase
         $this->getDm()->flush();
         $this->getDm()->clear();
 
-        $articleTitles = array_values($data);
         $locales = array_keys($data);
 
         foreach ($expectedPaths as $i => $expectedPath) {
-            $expectedLocale = $locales[$i];
+            $localeIndex = $i % count($locales);
+            $expectedLocale = $locales[$localeIndex];
 
             $route = $this->getDm()->find(null, $expectedPath);
 
-            $this->assertNotNull($route);
+            $this->assertNotNull($route, 'Route: '.$expectedPath);
             $this->assertInstanceOf('Symfony\Cmf\Bundle\RoutingAutoBundle\Model\AutoRoute', $route);
-            $this->assertEquals($expectedLocale, $route->getAutoRouteTag());
+            $this->assertEquals($expectedLocale, $route->getLocale());
 
             $content = $route->getContent();
 
@@ -271,12 +289,35 @@ class AutoRouteListenerTest extends BaseTestCase
         );
     }
 
+    public function testMultilangArticleRemainsSameLocale()
+    {
+        $article = new Article();
+        $article->path = '/test/article-1';
+        $article->title = 'Good Day';
+        $this->getDm()->persist($article);
+        $this->getDm()->flush();
+
+        $article->title = 'Hello everybody!';
+        $this->getDm()->bindTranslation($article, 'en');
+
+        $article->title = 'Bonjour le monde!';
+        $this->getDm()->bindTranslation($article, 'fr');
+
+        // let current article be something else than the last bound locale
+        $this->getDm()->findTranslation(get_class($article), $this->getDm()->getUnitOfWork()->getDocumentId($article), 'en');
+
+        $this->getDm()->flush();
+        $this->getDm()->clear();
+
+        $this->assertEquals('Hello everybody!', $article->title);
+    }
+
     /**
      * @dataProvider provideUpdateMultilangArticle
      */
     public function testUpdateMultilangArticle($data, $expectedPaths)
     {
-        $article = new Article;
+        $article = new Article();
         $article->path = '/test/article-1';
         $this->getDm()->persist($article);
 
@@ -296,11 +337,13 @@ class AutoRouteListenerTest extends BaseTestCase
 
         $article_de = $this->getDm()->findTranslation('Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document\Article', '/test/article-1', 'de');
         $routes = $this->getDm()->getReferrers($article_de);
-        $this->assertCount(count($data), $routes);
+
+        // Multiply the expected paths by 3 because Article has 3 routes defined.
+        $this->assertCount(count($data) * 3, $routes);
 
         $this->getDm()->clear();
 
-        foreach ($expectedPaths as $i => $expectedPath) {
+        foreach ($expectedPaths as $expectedPath) {
             $route = $this->getDm()->find(null, $expectedPath);
 
             $this->assertNotNull($route);
@@ -341,7 +384,7 @@ class AutoRouteListenerTest extends BaseTestCase
                 array(
                     'test/auto-route/seo-articles/en/goodbye-everybody',
                     'test/auto-route/seo-articles/fr/aurevoir-le-monde',
-                    'test/auto-route/seo-articles/de/aud-weidersehn',
+                    'test/auto-route/seo-articles/de/auf-weidersehn',
                     'test/auto-route/seo-articles/es/adios-todo-el-mundo',
                 ),
             ),
@@ -353,7 +396,7 @@ class AutoRouteListenerTest extends BaseTestCase
      */
     public function testLeaveRedirect($data, $updatedData, $expectedRedirectRoutePaths, $expectedAutoRoutePaths)
     {
-        $article = new SeoArticleMultilang;
+        $article = new SeoArticleMultilang();
         $article->title = 'Hai';
         $article->path = '/test/article-1';
         $this->getDm()->persist($article);
@@ -376,14 +419,14 @@ class AutoRouteListenerTest extends BaseTestCase
 
         foreach ($expectedRedirectRoutePaths as $originalPath) {
             $redirectRoute = $this->getDm()->find(null, $originalPath);
-            $this->assertNotNull($redirectRoute, 'Autoroute exists for: ' . $originalPath);
+            $this->assertNotNull($redirectRoute, 'Redirect exists for: '.$originalPath);
             $this->assertEquals(AutoRouteInterface::TYPE_REDIRECT, $redirectRoute->getDefault('type'));
         }
 
         foreach ($expectedAutoRoutePaths as $newPath) {
             $autoRoute = $this->getDm()->find(null, $newPath);
-            $this->assertNotNull($redirectRoute, 'Autoroute exists for: ' . $originalPath);
-            $this->assertEquals(AutoRouteInterface::TYPE_REDIRECT, $redirectRoute->getDefault('type'));
+            $this->assertNotNull($autoRoute, 'Autoroute exists for: '.$newPath);
+            $this->assertEquals(AutoRouteInterface::TYPE_PRIMARY, $autoRoute->getDefault('type'));
         }
     }
 
@@ -394,7 +437,7 @@ class AutoRouteListenerTest extends BaseTestCase
      */
     public function testLeaveRedirectAndRenameToOriginal()
     {
-        $article = new SeoArticle;
+        $article = new SeoArticle();
         $article->title = 'Hai';
         $article->path = '/test/article-1';
         $this->getDm()->persist($article);
@@ -410,7 +453,39 @@ class AutoRouteListenerTest extends BaseTestCase
     }
 
     /**
-     * Ensure that we can map parent classes: #56
+     * Leave direct should migrate children.
+     */
+    public function testLeaveRedirectChildrenMigrations()
+    {
+        $article1 = new SeoArticle();
+        $article1->title = 'Hai';
+        $article1->path = '/test/article-1';
+        $this->getDm()->persist($article1);
+        $this->getDm()->flush();
+
+        // add a child to the route
+        $parentRoute = $this->getDm()->find(null, '/test/auto-route/seo-articles/hai');
+        $childRoute = new AutoRoute();
+        $childRoute->setName('foo');
+        $childRoute->setParentDocument($parentRoute);
+        $this->getDm()->persist($childRoute);
+        $this->getDm()->flush();
+
+        $article1->title = 'Ho';
+        $this->getDm()->persist($article1);
+        $this->getDm()->flush();
+
+        $originalRoute = $this->getDm()->find(null, '/test/auto-route/seo-articles/hai');
+        $this->assertNotNull($originalRoute);
+        $this->assertCount(0, $this->getDm()->getChildren($originalRoute));
+
+        $newRoute = $this->getDm()->find(null, '/test/auto-route/seo-articles/ho');
+        $this->assertNotNull($newRoute);
+        $this->assertCount(1, $this->getDm()->getChildren($newRoute));
+    }
+
+    /**
+     * Ensure that we can map parent classes: #56.
      */
     public function testParentClassMapping()
     {
@@ -432,7 +507,7 @@ class AutoRouteListenerTest extends BaseTestCase
         $this->createBlog();
         $blog = $this->getDm()->find(null, '/test/test-blog');
 
-        $post = new Post;
+        $post = new Post();
         $post->name = 'Post 1';
         $post->title = 'Same Title';
         $post->blog = $blog;
@@ -440,7 +515,7 @@ class AutoRouteListenerTest extends BaseTestCase
         $this->getDm()->persist($post);
         $this->getDm()->flush();
 
-        $post = new Post;
+        $post = new Post();
         $post->name = 'Post 2';
         $post->title = 'Same Title';
         $post->blog = $blog;
@@ -448,7 +523,7 @@ class AutoRouteListenerTest extends BaseTestCase
         $this->getDm()->persist($post);
         $this->getDm()->flush();
 
-        $post = new Post;
+        $post = new Post();
         $post->name = 'Post 3';
         $post->title = 'Same Title';
         $post->blog = $blog;
@@ -470,7 +545,7 @@ class AutoRouteListenerTest extends BaseTestCase
 
     public function testCreationOfChildOnRoot()
     {
-        $page = new Page;
+        $page = new Page();
         $page->title = 'Home';
         $page->path = '/test/home';
         $this->getDm()->persist($page);
@@ -483,20 +558,53 @@ class AutoRouteListenerTest extends BaseTestCase
     }
 
     /**
-     * @expectedException Symfony\Cmf\Component\RoutingAuto\ConflictResolver\Exception\ExistingUriException
+     * @expectedException \Symfony\Cmf\Component\RoutingAuto\ConflictResolver\Exception\ExistingUriException
      */
     public function testConflictResolverDefaultThrowException()
     {
-        $blog = new Blog;
+        $blog = new Blog();
         $blog->path = '/test/test-blog';
         $blog->title = 'Unit testing blog';
         $this->getDm()->persist($blog);
         $this->getDm()->flush();
 
-        $blog = new Blog;
+        $blog = new Blog();
         $blog->path = '/test/test-blog-the-second';
         $blog->title = 'Unit testing blog';
         $this->getDm()->persist($blog);
         $this->getDm()->flush();
+    }
+
+    public function testGenericNodeShouldBeConvertedInAnAutoRouteNode()
+    {
+        $blog = new Blog();
+        $blog->path = '/test/my-post';
+        $blog->title = 'My Post';
+        $this->getDm()->persist($blog);
+        $this->getDm()->flush();
+
+        $this->assertInstanceOf(
+            'Doctrine\ODM\PHPCR\Document\Generic',
+            $this->getDm()->find(null, '/test/auto-route/blog')
+        );
+        $blogRoute = $this->getDm()->find(null, '/test/auto-route/blog/my-post');
+        $this->assertInstanceOf('Symfony\Cmf\Component\RoutingAuto\Model\AutoRouteInterface', $blogRoute);
+        $this->assertSame($blog, $blogRoute->getContent());
+
+        $page = new Page();
+        $page->path = '/test/blog';
+        $page->title = 'Blog';
+
+        $this->getDm()->persist($page);
+        $this->getDm()->flush();
+
+        $this->assertInstanceOf(
+            'Symfony\Cmf\Component\RoutingAuto\Model\AutoRouteInterface',
+            $this->getDm()->find(null, '/test/auto-route/blog')
+        );
+        $this->assertInstanceOf(
+            'Symfony\Cmf\Component\RoutingAuto\Model\AutoRouteInterface',
+            $this->getDm()->find(null, '/test/auto-route/blog/my-post')
+        );
     }
 }