+++ Tests/Resources/Document/Article.php
@@ -3,12 +3,11 @@
 /*
  * This file is part of the Symfony CMF package.
  *
- * (c) 2011-2014 Symfony CMF
+ * (c) 2011-2015 Symfony CMF
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 
 namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document;
 
@@ -33,7 +32,7 @@ class Article
     public $routes;
 
     /**
-     * @PHPCR\String(translated=true)
+     * @PHPCR\Field(type="string", translated=true)
      */
     public $title;
 
@@ -43,7 +42,7 @@ class Article
     public $locale;
 
     /**
-     * @PHPCR\Date(nullable=true)
+     * @PHPCR\Field(type="date", nullable=true)
      */
     public $date;
 
@@ -61,5 +60,4 @@ class Article
     {
         $this->date = $date;
     }
-
 }