+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 Changelog
 =========
 
+2.0.0-RC1
+---------
+
+* Added Symfony 3 support
+* [BC BREAK] Removed all `*.class` parameters.
+* Convert non-managed intermediate nodes into AutoRoute documents #165
+* Added support for multiple routes on the same document
+
+1.1.0
+-----
+
+Released.
+
+1.0.0
+-----
+
+* Enabled new `container` token provider - retrieve URL elements
+  from container parameters.
+
 1.0.0-RC1
 ---------